+++ src/compas_viewer/scene/scene.py
@@ -89,12 +89,11 @@ class ViewerScene(Scene):
         show_faces: Optional[bool] = None,
         pointcolor: Optional[Color] = None,
         linecolor: Optional[Color] = None,
-        surfacecolor: Optional[Color] = None,
         vertexcolor: Optional[Union[Color, dict[Any, Color]]] = None,
         edgecolor: Optional[Union[Color, dict[Any, Color]]] = None,
         facecolor: Optional[Union[Color, dict[Any, Color]]] = None,
-        lineswidth: Optional[float] = None,
-        pointssize: Optional[float] = None,
+        linewidth: Optional[float] = None,
+        pointsize: Optional[float] = None,
         opacity: Optional[float] = None,
         hide_coplanaredges: Optional[bool] = None,
         use_vertexcolors: Optional[bool] = None,
@@ -132,17 +131,15 @@ class ViewerScene(Scene):
             The single color of colors of the points in the COMPAS Geometry.
         linecolor : :class:`compas.colors.Color`, optional
             The single color of colors of the lines in the COMPAS Geometry.
-        surfacecolor : :class:`compas.colors.Color`, optional
-            The single color of colors the surfaces in the COMPAS Geometry.
         vertexcolor : Union[:class:`compas.colors.Color`, dict[Any, :class:`compas.colors.Color`], optional
             The color or the dict of colors of the vertices in the COMPAS Mesh.
         edgecolor : Union[:class:`compas.colors.Color`, dict[Any, :class:`compas.colors.Color`], optional
             The color or the dict of colors of the edges in the COMPAS Mesh.
         facecolor : Union[:class:`compas.colors.Color`, dict[Any, :class:`compas.colors.Color`], optional
             The color or the dict of colors of the faces in the COMPAS Mesh.
-        lineswidth : float, optional
+        linewidth : float, optional
             The line width to be drawn on screen
-        pointssize : float, optional
+        pointsize : float, optional
             The point size to be drawn on screen
         opacity : float, optional
             The opacity of the object.
@@ -175,11 +172,10 @@ class ViewerScene(Scene):
             pointcolor=pointcolor,
             linecolor=linecolor,
             facecolor=facecolor,
-            surfacecolor=surfacecolor,
             vertexcolor=vertexcolor,
             edgecolor=edgecolor,
-            lineswidth=lineswidth,
-            pointssize=pointssize,
+            linewidth=linewidth,
+            pointsize=pointsize,
             opacity=opacity,
             hide_coplanaredges=hide_coplanaredges,
             use_vertexcolors=use_vertexcolors,