+++ src/compas_viewer/ui/menubar.py
@@ -1,13 +1,13 @@
-from typing import TYPE_CHECKING
+from PySide6.QtGui import QAction
 
 from compas_viewer.base import Base
+from compas_viewer.components.combobox import ViewModeAction
+from compas_viewer.components.dialog import CameraSettingsDialog
 
-if TYPE_CHECKING:
-    pass
 
-
-def new_file():
-    print("new file...")
+def openDialog():
+    dialog = CameraSettingsDialog()
+    dialog.exec()
 
 
 class MenuBar(Base):
@@ -16,5 +16,17 @@ class MenuBar(Base):
 
     def lazy_init(self):
         self.widget = self.viewer.ui.window.menuBar()
-        filemenu = self.widget.addMenu("File")
-        filemenu.addAction("New File...", new_file)
+
+        camera_filemenu = self.widget.addMenu("Camera")
+
+        camera_filemenu.addAction("Camera_Settings", openDialog)
+        viewmode_menu = camera_filemenu.addMenu("Viewmode")
+        viewmode_menu.addAction(self.viewmode_action("perspective"))
+        viewmode_menu.addAction(self.viewmode_action("top"))
+        viewmode_menu.addAction(self.viewmode_action("front"))
+        viewmode_menu.addAction(self.viewmode_action("right"))
+
+    def viewmode_action(self, mode: str):
+        action = QAction(mode, self.widget)
+        action.triggered.connect(lambda check=False, mode=mode: ViewModeAction().change_view(mode))
+        return action