+++ src/compas_viewer/ui/toolbar.py
@@ -1,9 +1,12 @@
 from compas_viewer.base import Base
-from compas_viewer.components.button_factory import ButtonFactory
+from compas_viewer.components.button import Button
+from compas_viewer.components.combobox import ViewModeAction
+from compas_viewer.components.dialog import CameraSettingsDialog
 
 
-def test_action() -> None:
-    print("test action...")
+def openDialog():
+    dialog = CameraSettingsDialog()
+    dialog.exec()
 
 
 class ToolBar(Base):
@@ -15,4 +18,5 @@ class ToolBar(Base):
         self.widget.setMovable(False)
         self.widget.setObjectName("Tools")
         self.widget.setHidden(not self.viewer.config.ui.toolbar.show)
-        self.widget.addWidget(ButtonFactory("zoom_selected.svg", "zoom", test_action()))
+        self.widget.addWidget(ViewModeAction().combobox())
+        self.widget.addWidget(Button("camera_info.svg", "Camera_Settings", openDialog))