+++ src/compas_viewer/viewer.py
@@ -1,7 +1,13 @@
+import os
 import sys
+from typing import Callable
+from typing import Optional
 
+from PySide6.QtCore import QTimer
+from PySide6.QtGui import QIcon
 from PySide6.QtWidgets import QApplication
 
+from compas_viewer import HERE
 from compas_viewer.components.renderer import Renderer
 from compas_viewer.config import Config
 from compas_viewer.configurations import ControllerConfig
@@ -16,6 +22,8 @@ from compas_viewer.view3d.view3d import View3D
 class Viewer(Singleton):
     def __init__(self, *args, **kwargs):
         self.app = QApplication(sys.argv)
+        self.app.setWindowIcon(QIcon(os.path.join(HERE, "icons", "compas_icon_white.png")))
+        self.timer = QTimer()
         self.config = Config()
         self.scene = ViewerScene()
         
@@ -29,3 +37,34 @@ class Viewer(Singleton):
         self.ui.lazy_init()
         self.ui.show()
         self.app.exec()
+
+    def on(self, interval: int, frames: Optional[int] = None) -> Callable:
+        """Decorator for callbacks of a dynamic drawing process with fixed intervals.
+
+        Parameters
+        interval : int
+            Interval between subsequent calls to this function, in milliseconds.
+        frames : int, optional
+            The number of frames of the process.
+            If no frame number is provided, the process continues until the viewer is closed.
+
+        Returns
+        -------
+        Callable
+        """
+        self.frame_count = 0
+
+        def decorator(func: Callable):
+            def wrapper():
+                if frames is not None and self.frame_count >= frames:
+                    self.timer.stop()
+                    return
+                func(self.frame_count)
+                self.frame_count += 1
+                self.renderer.update()
+
+            self.timer.timeout.connect(wrapper)
+            self.timer.start(interval)
+            return func
+
+        return decorator